Add web-audio-test-api
The Web Audio API is not available in Node which is required for almost
all of this libraries code to run. The package ‘web-audio-test-api’
basically mocks all of the functions of this browser API. The library
hasn’t been updated in a few years which means we may find that in the
near future we need to fork and update it ourselves.