Migrate from web components to vue
Vue handles a lot of things automatically, provides logic and loops in the HTML template string, and has some cool syntax for automatically binding object properties to the main synth instance. Ultimately the UI is exactly the same but it has better browser support and the overall syntax is cleaner and will be easier to extend in the future.